LayoutTests/imported/w3c/web-platform-tests/background-fetch/interfaces-worker.https-expected.txt

CONSOLE MESSAGE: line 153: TypeError: undefined is not an object (evaluating 'registration.unregister')
idlharness test

This test validates the WebIDL included in the Background Fetch API (Service Workers).


FAIL Service Worker-scoped tests. assert_unreached: unregister should not fail: serviceWorker.getRegistration() is not yet implemented Reached unreachable code
